requirements.txt
django
djangorestframework
django-cors-headers
Pillow
mysqlclient
